requirements.txt
fastapi
uvicorn
torch
numpy
alpha_vantage
python-dotenv
mangum